1. Added unit tests for add and delete 2. Fixed the bug where delete request is done more than once by keeping an array in user defaults for unsynced deletes, similarly for unsynced updates, and this array is separate for the prior array that wrote all deletes to user defaults whether synced or not. That array is no longer written to user defaults, similarly for all updates synced or not.
The purpose of keeping track of all updates and deletions whether synced or not, is to get around an issue where if a device has been updated (checked in/out) and then deleted, deletion will not take affect until the app is restarted. Keeping track of all updates and deletions, the deleted devices can be removed from the tableview even if the deletion hasn't taken effect in Core Data. I initially used the same array to keep track of unsynced changes, resulting in multiple deletions in some cases, since the array was written to whether the app was offline or online, and once it went online it will redo the deletions, similarly for updates.